Switched to NSInteger

+++ OTPClient/ObjectMapping/OTPObjectManager.h
@@ -23,11 +23,11 @@ typedef void (^OTPTripPlanCompletionHandler)(OTPTripPlan *tripPlan, NSError *err
 - (void)loadTripPlanFrom:(CLLocationCoordinate2D)from
                       to:(CLLocationCoordinate2D)to
                     date:(NSDate *)date
-          numItineraries:(int)numItineraries
+          numItineraries:(NSInteger)numItineraries
           shouldArriveBy:(BOOL)shouldArriveBy
    requiresAccessibility:(BOOL)requiresAccessibility
-         maxWalkDistance:(int)maxWalkDistance
-         transferPenalty:(int)transferPenalty
+         maxWalkDistance:(NSInteger)maxWalkDistance
+         transferPenalty:(NSInteger)transferPenalty
        completionHandler:(OTPTripPlanCompletionHandler)completionHandler;
 
 - (void)loadTripPlanFrom:(CLLocationCoordinate2D)from
